GEODE-5746: Implement Cluster class
- higher-level abstraction for test classes that don't need fine control

+++ clicache/integration-test2/Gfsh.cs
@@ -133,7 +133,7 @@ namespace Apache.Geode.Client.IntegrationTests
                     gfsh_.LocatorPort = port;
                     return this;
                 }
-                public Locator withJmxManagerPort(short jmxManagerPort)
+                public Locator withJmxManagerPort(int jmxManagerPort)
                 {
                     gfsh_.JmxManagerPort = jmxManagerPort;
                     return this;
